+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - 2026.04.15
 
   - Fix - IBM DB2 z/OS Connector: Allow "@" as a statement terminator in addition to ";". Added debug messages to trace the DDL parsing process.
+  - Fix - T-SQL Parser: Fixed missing comments in PostgreSQL converted code, fixed conversion of data types in stored procedures, key word "noholdlock" removed from select statements.
 
 - 2026.03.30